% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoricEvSyn.r
\name{GoricEvSyn}
\alias{GoricEvSyn}
\title{GORIC(A) evidence synthesis}
\usage{
GoricEvSyn(
  TypeEv,
  S,
  Param_studies,
  CovMx_studies,
  SameHypo,
  NrHypos,
  Hypo_studies,
  ComplOrUnc = 1
)
}
\arguments{
\item{TypeEv}{The type of evidence-synthesis approach: Equal-evidence approach (0) or Added-evidence approach (1).
In case of an equal-evidence approach, aggregating evidence from, say, 5 studies with n=100 observations is the same as obtaining evidence from 1 study (as if it was possible) with n=500 observations (like meta-analysis does).
In the added-evidence approach, the aggregated evidence from, says, 5 studies is stronger than as if the data were combined (as if that was possible).}

\item{S}{The number of (primary) studies. That is, the results (evidence) of S studies will be aggregated.}

\item{Param_studies}{List of S 'named' vectors with the (standardized) parameter estimates of interest. Thus, there are S items in the list and each item is a 'named' vector with the number of parameter estimates relevant for that study. Note: The names of the estimates should be used in the hypothesis specification.}

\item{CovMx_studies}{List of the S covariance matrices of the (standardized) parameter estimates of interest.}

\item{SameHypo}{Indicator whether the same hypotheses are used (1) or not (0). If SameHypo = 1, then the same estimates in Param_studies should have the same name.}

\item{NrHypos}{The number of theory-based hypotheses that will be evaluated.}

\item{Hypo_studies}{The NrHypos theory-based hypotheses. If SameHypo = 0, then there should be S specifications of the NrHypos theory-based hypotheses.}

\item{ComplOrUnc}{Indicator of which safeguard-hypothesis should be used in case 'NrHypos = 1': Complement (0; i.e., the remaining theories) or unconstrained (1; i.e., all possible theories including the one specified). If NrHypos > 1, then automatically the unconstrained will be used.}
}
\value{
The output comprises, among others, the overall evidence for the theory-based hypotheses.
}
\description{
GORIC(A) evidence synthesis (GoricEvSyn) aggregates the evidence for theory-based hypotheses from multiple studies that may use diverse designs to investigate the same central theory. There is also an interactive web application on my website to perform GoricEvSyn: \url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}.
}
\examples{

# In progress
}
