% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoricEvSyn_IC.r
\name{GoricEvSyn_IC}
\alias{GoricEvSyn_IC}
\title{GORIC(A) evidence synthesis based on AIC or ORIC or GORIC or GORICA values}
\usage{
GoricEvSyn_IC(S, IC)
}
\arguments{
\item{S}{The number of (primary) studies. That is, the results (evidence) of S studies will be aggregated.}

\item{IC}{A matrix with information criteria (AIC, ORIC, GORIC, or GORICA) values of size S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained).}
}
\value{
The output comprises, among others, the overall evidence for the theory-based hypotheses.
}
\description{
GORIC(A) evidence synthesis (GoricEvSyn) aggregates the evidence for theory-based hypotheses from multiple studies that may use diverse designs to investigate the same central theory. There is also an interactive web application on my website to perform GoricEvSyn: \url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}.
In case IC values are used as input, the added-evidence approach is used in which the aggregated evidence from, says, 5 studies is stronger than as if the data were combined (as if that was possible).
}
\examples{

S <- 4
IC <- myGORICs # Example based on S = 4 studies and 3 hypotheses:
# H0 <- "beta1 == 0"
# Hpos <- "beta1 > 0"
# Hneg <- "beta1 < 0"
# Note that in this set the whole space is (all theories are) covered so the unconstrained is not needed as safeguard-hypothesis
GoricEvSyn_IC(S, IC)
}
