% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoricEvSyn_LLandPT.r
\name{GoricEvSyn_LLandPT}
\alias{GoricEvSyn_LLandPT}
\title{GORIC(A) evidence synthesis based on log likelihood and penalty values}
\usage{
GoricEvSyn_LLandPT(TypeEv, S, LL, PT)
}
\arguments{
\item{TypeEv}{The type of evidence-synthesis approach: Equal-evidence approach (0) or Added-evidence approach (1).
In case of an equal-evidence approach, aggregating evidence from, say, 5 studies with n=100 observations is the same as obtaining evidence from 1 study (as if it was possible) with n=500 observations (like meta-analysis does).
In the added-evidence approach, the aggregated evidence from, says, 5 studies is stronger than as if the data were combined (as if that was possible).}

\item{S}{The number of (primary) studies. That is, the results (evidence) of S studies will be aggregated.}

\item{LL}{A matrix with log likelihood values of size S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained).}

\item{PT}{A matrix with penalty values of size S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained).}
}
\value{
The output comprises, among others, the overall evidence for the theory-based hypotheses.
}
\description{
GORIC(A) evidence synthesis (GoricEvSyn) aggregates the evidence for theory-based hypotheses from multiple studies that may use diverse designs to investigate the same central theory. There is also an interactive web application on my website to perform GoricEvSyn: \url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}.
}
\examples{

S <- 4
# Example based on S = 4 studies and 3 hypotheses:
# H0 <- "beta1 == 0"
# Hpos <- "beta1 > 0"
# Hneg <- "beta1 < 0"
# Note that in this set the whole space is (all theories are) covered so the unconstrained is not needed as safeguard-hypothesis
LL <- myLLs
PT <- myPTs

# Added-evidence approach
TypeEv <- 1
GoricEvSyn_LLandPT(TypeEv, S, LL, PT)

# Equal-evidence approach
TypeEv <- 0
GoricEvSyn_LLandPT(TypeEv, S, LL, PT)
}
