% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBFs.R, R/myWeights.R
\docType{data}
\name{myBFs}
\alias{myBFs}
\title{An example of Bayes factor (BF) values}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 3 columns.
}
\usage{
data(myBFs)

data(myBFs)
}
\description{
An example of BF: A matrix with BF values of size 4 x 3 (in general: S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained)).

An example of weights: A matrix with information criteria (AIC, ORIC, GORIC, or GORICA) weights of size 4 x 3 (in general: S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained)).
}
\examples{
data(myBFs)

data(myBFs)

}
\keyword{datasets}
