% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPTs.R
\docType{data}
\name{myPTs}
\alias{myPTs}
\title{An example of penalty (PT) values}
\format{
An object of class \code{matrix} with 4 rows and 3 columns.
}
\usage{
data(myPTs)
}
\description{
An example of PT: A matrix with penalty values of size 4 x 3 (in general: S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained)).
}
\examples{
data(myPTs)

}
\keyword{datasets}
