% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myWeights.R
\docType{data}
\name{myWeights}
\alias{myWeights}
\title{An example of GORIC weight values}
\format{An object of class \code{matrix} with 4 rows and 3 columns.}
\usage{
data(myWeights)
}
\description{
An example of weights: A matrix with information criteria (AIC, ORIC, GORIC, or GORICA) weights of size 4 x 3 (in general: S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained)).
}
\examples{
data(myWeights)

}
\keyword{datasets}
