% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoricEvSyn_IC.r
\name{GoricEvSyn_IC}
\alias{GoricEvSyn_IC}
\title{GORIC(A) evidence synthesis based on AIC or ORIC or GORIC or GORICA values}
\usage{
GoricEvSyn_IC(S, IC, Name_studies = 1:S, PrintPlot = T)
}
\arguments{
\item{S}{The number of (primary) studies. That is, the results (evidence) of S studies will be aggregated.}

\item{IC}{A matrix with information criteria (AIC, ORIC, GORIC, or GORICA) values of size S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained).}

\item{Name_studies}{Vector of S numbers or S characters to be printed at the x-axis of the plot with GORIC(A) weights. Default: Name_studies = 1:S.}

\item{PrintPlot}{Indicator whether plot of GORIC(A) weigths should be printed (TRUE; default) or not (FALSE). The GORIC(A) weights per study are plotted and the cumulative GORIC(A) weights (where those for last study are the final ones).
#'}
}
\value{
The output comprises, among other things, the cumulative and final evidence for the theory-based hypotheses.
}
\description{
GORIC(A) evidence synthesis (GoricEvSyn) aggregates the evidence for theory-based hypotheses from multiple studies that may use diverse designs to investigate the same central theory. There is also an interactive web application on my website to perform GoricEvSyn: \url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}.
In case IC values are used as input, the added-evidence approach is used in which the aggregated evidence from, says, 5 studies is stronger than as if the data were combined (as if that was possible).
}
\examples{

S <- 4
IC <- myGORICs # Example based on S = 4 studies and 3 hypotheses:
# H0 <- "beta1 == 0"  # this hypothesis could have been left out
# Hpos <- "beta1 > 0"
# Hneg <- "beta1 < 0"
# Note that in this set the whole space is (all theories are) covered so the unconstrained is not needed as safeguard-hypothesis
GoricEvSyn_IC(S, IC)

# Change labels on x-axis in GORIC(A) weigths plot #
# For example, let us say that the studies come from the years 2015, 2016, 2017, 2019.
# Because of unequal spacing, you may want to use numbers instead of characters:
Name_studies <- c(2015, 2016, 2017, 2019)
GoricEvSyn_IC(S, IC, Name_studies)
}
